Refactor agent tool registry into modules
